migrate: move vbaFacility to products directory
- Migrated vbaFacility product files to src/products/vbaFacility/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved additional component: vbaFacilityServiceGroup.tsx and its test
- Moved query.ts from src/data/queries/ to src/products/vbaFacility/
- Moved test file and renamed vbaFacility.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/vbaFacility/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/vbaFacility/mock.json
- Updated all import statements across the codebase
- Fixed import paths in migrated files (template.test.tsx, query.test.ts)
- Moved test snapshot file to new location
- All TypeScript errors resolved
- All tests passing (3 query tests + 5 template tests + 1 component test = 9 total)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and vbaFacility pages load correctly
