+++ src/products/vbaFacility/template.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import { VbaFacility } from './index'
-import mockData from '@/mocks/vbaFacility.mock.json'
+import { VbaFacility } from './template'
+import mockData from '@/products/vbaFacility/mock.json'
 import mockServiceData from '@/mocks/vbaFacilityService.mock.json'
-import { formatter } from '@/data/queries/vbaFacility'
+import { formatter } from '@/products/vbaFacility/query'
 
 describe('VbaFacility with valid data', () => {
   const mockWithService = {
